core: transport: Add I2C and UART Notecard transports
As the drivers implementing the device-tree instances.
The I2C transport initializes and provides the device-hooks required
for the notecard middleware to interact with the device.
The UART transport is not yet implemented and enabling it shall throw
a build-time failure with a descriptive error message.

Both implementations only support single-instance driver, and they're
mutually exclusive: either UART or I2C can be enabled at a time,
otherwise the build results in an error: this given the hooks APIs
don't provide a way to pass on an object/struct instance to reference
back to (e.g: device reference).

Signed-off-by: Luis Ubieda <luisf@croxel.com>